--- src/main/java/com/study/lecture/presentation/LectureDTO.java
@@ -1,4 +1,0 @@
-package com.study.lecture.presentation;
-
-public class LectureDTO {
-}
