+++ src/main/java/com/study/lecture/presentation/dto/request/LectureRequestDTO.java
@@ -1,0 +1,4 @@
+package com.study.lecture.presentation.dto.request;
+
+public record LectureRequestDTO() {
+}
