+++ src/main/java/com/study/lecture/presentation/dto/response/LectureResponseDTO.java
@@ -1,0 +1,4 @@
+package com.study.lecture.presentation.dto.response;
+
+public record LectureResponseDTO() {
+}
